requirements.txt
streamlit
gspread==5.11.3
pandas
googletrans==4.0.0-rc1
PyMuPDF
Pillow
hijri-converter
httpcore==0.9.1
httpx==0.13.3
persiantools==3.0.1
google-cloud
